Add typescript support
