+++ .gitignore
@@ -3,3 +3,5 @@ npm-debug.log
 .env
 .tmp
 .DS_Store
+lib/
+conda.txt
