+++ .prettierrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+    semi: true,
+    trailingComma: 'es5',
+    singleQuote: true,
+    printWidth: 120,
+    tabWidth: 4,
+    useTabs: false,
+};
