--- lib/util.js
@@ -1,5 +1,0 @@
-import {URL} from 'url';
-
-export function absUrl(url, base) {
-  return new URL(url, base).href;
-}
