+++ src/util.ts
@@ -1,0 +1,5 @@
+import { URL } from 'url';
+
+export function absUrl(url: string, base: string): string {
+    return new URL(url, base).href;
+}
