+++ typings/microformat-node.d.ts
@@ -1,0 +1,18 @@
+import { MicroformatData } from "../src/types";
+
+declare module 'microformat-node' {
+
+    declare namespace MicroformatNode {
+        interface Root {
+            get(options: GetOptions, cb: ReqCallback): void;
+        }
+
+        interface GetOptions {
+            html: string;
+        }
+        type ReqCallback = (err: any, data: MicroformatData) => void;
+    }
+
+    const module: MicroformatNode.Root;
+    export = module;
+};
